% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_metrics.R
\name{align_metrics}
\alias{align_metrics}
\title{Extract and format cleaning and alignment metrics}
\usage{
align_metrics(
  data.dir = NULL,
  trim = TRUE,
  bam = TRUE,
  picard = TRUE,
  bam.filter = TRUE,
  count = TRUE
)
}
\arguments{
\item{data.dir}{Character string of directory containing all associated files}

\item{trim}{Logical if should include FastQC trim settings}

\item{bam}{Logical if should include samtools flagstat for raw alignments}

\item{picard}{Logical if should include Picard for raw alignments}

\item{bam.filter}{Logical if should include samtools flagstat for filtered alignments}

\item{count}{Logical if should include featureCounts total reads in genes}
}
\value{
Data frame cleaning and alignment metrics for all libraries
}
\description{
Extract data from FastQC trim settings, Picard, samtools flagstat, and featureCounts output by the RNA-seq fastq pipeline
}
