% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lmFit.R
\name{extract_lmFit}
\alias{extract_lmFit}
\title{Extract lmFit model results}
\usage{
extract_lmFit(
  design,
  fit,
  contrast.mat = NULL,
  dat.genes = NULL,
  name.genes = "geneName"
)
}
\arguments{
\item{design}{model matrix output by model.matrix( )}

\item{fit}{MArrayLM model fit output by limma::eBayes( )}

\item{contrast.mat}{contrast matrix output by limma::makeContrasts( )}

\item{dat.genes}{data frame with additional gene annotations. Optional. Default is NULL}

\item{name.genes}{character for variable name in dat.genes that matches gene names in fit. Default is "geneName"}
}
\value{
Data frame with model fit and significant for all variable and genes
}
\description{
Extract model fit and significance for all individual variables and/or contrasts
}
\examples{
# Run limma model
design <- model.matrix(~ virus, data = dat.voom.example$targets)
fit <- limma::eBayes(limma::lmFit(dat.voom.example$E, design))

## Get results
result <- extract_lmFit(design = design, fit = fit)
## Get results and add gene annotations
fdr <- extract_lmFit(design = design, fit = fit,
                        dat.genes = dat.voom.example$genes, name.genes = "geneName")

# Run limma contrasts model
design <- model.matrix(~ 0 + virus, data = dat.voom.example$targets)
fit <- limma::lmFit(dat.voom.example$E, design)
contrast.mat <- limma::makeContrasts(virusB-virusA, levels = design)
fit <- eBayes(contrasts.fit(fit, contrast.mat))

## Get contrast results
fdr <- extract_lmFit(design = design, fit = fit, contrast.mat = contrast.mat)
}
