% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_plots.R
\name{gene_plots}
\alias{gene_plots}
\title{Title}
\usage{
gene_plots(
  dat = NULL,
  counts = NULL,
  meta = NULL,
  genes = NULL,
  fdr,
  model.name = NULL,
  libraryID = "libID",
  geneID = "geneName",
  subset.genes = NULL,
  variables,
  interaction = NULL,
  colorID = NULL,
  colors = NULL,
  ylab = "Normalized log2 expression",
  width = 5,
  height = 5,
  outdir = "figs/",
  cores = 2
)
}
\arguments{
\item{dat}{DGEList or EList object with expression data (counts, E), sample metadata (samples, targets), and gene annotation (genes)}

\item{counts}{If dat not provided. Data frame of gene expression. Genes are rows, samples are columns. geneID must be rownames or first column}

\item{meta}{If dat not provided. Data frame of sample meta data with samples as rows.}

\item{genes}{Optional if dat not provided. Data frame of gene annotation with genes as rows.}

\item{fdr}{Optional. Model results output by kmFit( )}

\item{model.name}{Optional. Character string of model name to include in fdr results}

\item{libraryID}{Character string of variable name to use when combining expression and sample data}

\item{geneID}{Character string of variable name to use when combining expression and gene data}

\item{subset.genes}{Optional. Character vector of genes to plot. Must match names in geneID column. If not provided, all genes are plotted.}

\item{variables}{Character vector of variable names to include in plot. Variables can be character, factor, or numeric}

\item{interaction}{Logical if plot interaction effect of first two variables}

\item{colorID}{Character string of variable to color data points}

\item{colors}{Optional character vector of colors for colorID. If not provided, default ggplot is used}

\item{ylab}{Character string of y-axis lab}

\item{width}{Numeric figure width in inches}

\item{height}{Numeric figure height in inches}

\item{outdir}{Character string of output directory}

\item{cores}{Numeric cores to run in parallel}
}
\value{
Save individual pdf plots for each gene to outdir
}
\description{
Title
}
\examples{
subset.genes <- c("ENSG00000250479","ENSG00000250510","ENSG00000255823")
fdr <- kmFit(dat = dat.voom.example,
      patientID = "donorID",
      kin = kin.example,
      subset.genes = subset.genes,
      model = "~ virus + (1|donorID)")
gene_plots(dat = dat.voom.example, fdr = fdr, subset.genes = subset.genes,
     variables = "virus", colorID = "virus")
}
