% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kimma_lme.R
\name{kimma_lme}
\alias{kimma_lme}
\title{Run kimma linear mixed effects model}
\usage{
kimma_lme(model.lme, to.model.gene, gene, use.weights)
}
\arguments{
\item{model.lme}{Character model created in kmFit}

\item{to.model.gene}{Data frame formatted in kmFit, subset to gene of interest}

\item{gene}{Character of gene to model}

\item{use.weights}{Logical if gene specific weights should be used in model. Default is FALSE}
}
\value{
Linear model effect results data frame for 1 gene
}
\description{
Run kimma linear mixed effects model
}
\keyword{internal}
