% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kimma_lmekin.R
\name{kimma_lmekin}
\alias{kimma_lmekin}
\title{Run kimma linear mixed effects model with kinship}
\usage{
kimma_lmekin(model.lme, to.model.gene, gene, kin.subset, use.weights)
}
\arguments{
\item{model.lme}{Character model created in kmFit}

\item{to.model.gene}{Data frame formatted in kmFit, subset to gene of interest}

\item{gene}{Character of gene to model}

\item{kin.subset}{Pairwise kinship matrix}

\item{use.weights}{Logical if gene specific weights should be used in model. Default is FALSE}
}
\value{
Linear model effect with kinship results data frame for 1 gene
}
\description{
Run kimma linear mixed effects model with kinship
}
\keyword{internal}
