% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kimma_lmerel.R
\name{kimma_lmerel}
\alias{kimma_lmerel}
\title{Run kimma linear mixed effects model with kinship}
\usage{
kimma_lmerel(
  model.lme,
  to.model.gene,
  gene,
  kin.subset,
  use.weights,
  patientID
)
}
\arguments{
\item{model.lme}{Character model created in kmFit}

\item{to.model.gene}{Data frame formatted in kmFit, subset to gene of interest}

\item{gene}{Character of gene to model}

\item{kin.subset}{Pairwise kinship matrix}

\item{use.weights}{Logical if gene specific weights should be used in model. Default is FALSE}

\item{patientID}{Character of variable name to match dat$targets to kinship row and column names.}
}
\description{
Run kimma linear mixed effects model with kinship
}
