% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin.example.R
\docType{data}
\name{kin.example}
\alias{kin.example}
\title{kimma example kinship.}
\format{
A matrix with 3 rows and 3 variables:
\describe{
  \item{rowname}{Donor ID. Same as column names}
  \item{donor1}{numeric kinship (0-1) with donor 1}
  \item{donor2}{numeric kinship (0-1) with donor 2}
  \item{donor3}{numeric kinship (0-1) with donor 3}
}
}
\usage{
kin.example
}
\description{
Matrix of pairwise kinship values between donor 1,2,3. Values are dummy data with 1 for self comparison, 0.5 for siblings, and 0.1 for unrelated.
}
\keyword{datasets}
