% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmFit_contrast.R
\name{kmFit_contrast}
\alias{kmFit_contrast}
\title{Run pairwise model comparisons with emmeans}
\usage{
kmFit_contrast(fit, contrast_var, to_model_gene, genotype_name)
}
\arguments{
\item{fit}{model fit from lm( ) or lmer( )}

\item{contrast_var}{Character vector of variable in model to run contrasts of}

\item{to_model_gene}{Formatted data from kimma_cleaning( ), subset to gene of interest}

\item{genotype_name}{Character string. Used internally for kmFit_eQTL}
}
\value{
data frame with contrast model results
}
\description{
Run pairwise model comparisons with emmeans
}
\keyword{internal}
