% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmFit_contrast_kin.R
\name{kmFit_contrast_kin}
\alias{kmFit_contrast_kin}
\title{Run pairwise model comparisons within lmekin model}
\usage{
kmFit_contrast_kin(
  contrast.var,
  to.model.gene,
  patientID,
  to.model.ls,
  gene,
  use.weights
)
}
\arguments{
\item{contrast.var}{Character vector of variable in model to run contrasts of}

\item{to.model.gene}{Formatted data from kimma_cleaning( ), subset to gene of interest}

\item{patientID}{Character of variable name to match dat$targets to kinship row and column names.}

\item{to.model.ls}{Formatted data from kimma_cleaning( )}

\item{gene}{Character name of gene of interest}
}
\value{
data frame with contrast model results
}
\description{
Run pairwise model comparisons within lmekin model
}
\keyword{internal}
