% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmFit_lm.R
\name{kmFit_lm}
\alias{kmFit_lm}
\title{Linear mixed effects models with kinship for RNA-seq}
\usage{
kmFit_lm(
  dat = NULL,
  kin = NULL,
  patientID = "ptID",
  libraryID = "libID",
  counts = NULL,
  meta = NULL,
  genes = NULL,
  weights = NULL,
  subset_var = NULL,
  subset_lvl = NULL,
  subset_genes = NULL,
  model,
  use_weights = FALSE,
  metrics = FALSE,
  run_contrast = FALSE,
  contrast_var = NULL,
  processors = NULL,
  p_method = "BH",
  genotype_name = NULL
)
}
\arguments{
\item{dat}{EList object output by voom( ). Must contain counts (dat$E) and meta (dat$targets). Optionally also contains gene metadata (dat$genes) and weights (dat$weights)}

\item{kin}{Matrix with pairwise kinship values between individuals. Must be numeric with rownames.}

\item{patientID}{Character of variable name to match dat$targets to kinship row and column names.}

\item{libraryID}{Character of variable name to match dat$targets to dat$E colnames}

\item{counts}{Matrix of normalized expression. Rows are genes, columns are libraries.}

\item{meta}{Matrix or data frame of sample and individual metadata.}

\item{genes}{Optional matrix or data frame of gene metadata.}

\item{weights}{Optional matrix of data frame of gene specific weights. Usually calculated with limma::voomWithQualityWeights().}

\item{subset_var}{Character list of variable name(s) to filter data by.}

\item{subset_lvl}{Character list of variable value(s) or level(s) to filter data to. Must match order of subset_var}

\item{subset_genes}{Character vector of genes to include in models.}

\item{model}{Character vector of model starting with ~ Should include (1|patientID) if mixed effects will be run}

\item{use_weights}{Logical if gene specific weights should be used in model. Default is FALSE}

\item{metrics}{Logical if should calculate model fit metrics such as AIC, BIC, R-squared. Default is FALSE}

\item{run_contrast}{Logical if should run pairwise contrasts. If no matrix provided, all possible pairwise comparisons are completed.}

\item{contrast_var}{Character vector of variable in model to run contrasts of. Interaction terms must be specified as "var1:var2". If NULL (default), all contrasts for all variables in the model are run}

\item{processors}{Numeric processors to run in parallel. Default is 2 less than the total available}

\item{p_method}{Character of FDR adjustment method. Values as in p.adjust( )}

\item{genotype_name}{Character string. Used internally for kmFit_eQTL}
}
\value{
List of data frames including
   - lm: model estimates and significance
   - *.contrast: model estimates and significance for pairwise contrasts with variables in the original model
   - *.fit: model fit metrics such as sigma, AIC, BIC, R-squared (optional with metrics paramater)
   - *.error: error messages for genes that failed model fitting
}
\description{
Run lmerel and corresponding lm or lme without kinship of gene expression in RNA-seq data
}
\examples{
#NEED TO BE UPDATED

}
