% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmekin2.R
\name{lmekin2}
\alias{lmekin2}
\title{Modification of coxme::lmekin( ) to use input weights}
\usage{
lmekin2(
  formula,
  data,
  weights,
  subset,
  na.action,
  control,
  varlist,
  vfixed,
  vinit,
  method = c("ML", "REML"),
  x = FALSE,
  y = FALSE,
  model = FALSE,
  random,
  fixed,
  variance,
  ...
)
}
\value{
Linear model effect with kinship
}
\description{
Modification of coxme::lmekin( ) to use input weights
}
\keyword{internal}
