% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_modules.R
\name{make_modules}
\alias{make_modules}
\title{Construct WGCNA modules and associated data}
\usage{
make_modules(
  dat,
  genes = NULL,
  Rsq.min = NULL,
  sft.value = NULL,
  minModuleSize = 20,
  deepSplit = 3,
  nThread = 2
)
}
\arguments{
\item{dat}{limma EList output by voom( )}

\item{genes}{Character vector of genes to used in module building. Must match rownames in dat. If not set, all genes in dat are used}

\item{Rsq.min}{Numeric minimum R-squared for soft threshold selection. If set, sft.value is not used}

\item{sft.value}{Numeric soft threshold. Set when minimum R-squared is no used}

\item{minModuleSize}{Numeric minimum module size}

\item{deepSplit}{Integer value between 0 and 4. Provides a simplified control over how sensitive module detection should be to module splitting, with 0 least and 4 most sensitive}

\item{nThread}{Integer for number of threads to use}
}
\value{
List including:
\itemize{
  \item{genes} Character vector of genes used in module building
  \item{sft} Data frame with soft thresholding selected for module building. Includes power, minimum R-squared, and connectivity
  \item{sft.plot} ggplot object of soft thresholding topology and connectivity
  \item{mods} Data frame of genes in modules
  \item{mods.voom} Data frame of mean module expression in each library
  \item{david} DAVID formatted data frame of genes in modules
}
}
\description{
Make WGCNA modules from gene expression data with dynamic soft threshold selection. Also outputs mean module expression and DAVID formatted gene lists
}
\examples{
dat.mods <- make_modules(dat = dat.voom.example, sft.value = 1)
}
