% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_kmFit.R
\name{summarise_kmFit}
\alias{summarise_kmFit}
\alias{summarize_kmFit}
\title{Summarise kmFit FDR results}
\usage{
summarise_kmFit(
  fdr,
  fdr_cutoff = c(0.05, 0.1, 0.2, 0.3, 0.4, 0.5),
  p_cutoff = NULL,
  FCgroup = FALSE,
  intercept = FALSE,
  fdr.cutoff = NULL,
  p.cutoff = NULL
)

summarize_kmFit(
  fdr,
  fdr_cutoff = c(0.05, 0.1, 0.2, 0.3, 0.4, 0.5),
  p_cutoff = NULL,
  FCgroup = FALSE,
  intercept = FALSE,
  fdr.cutoff = NULL,
  p.cutoff = NULL
)
}
\arguments{
\item{fdr}{data.frame output by kimma::kmFit( ). Main model or contrasts accepted}

\item{fdr_cutoff}{numeric vector of FDR cutoffs to summarise at}

\item{p_cutoff}{numeric vector of P-value cutoffs to summarise at. No FDR summary given if p.cutoff is provided}

\item{FCgroup}{logical if should separate summary by up/down fold change groups}

\item{intercept}{logical if should include intercept variable in summary}

\item{fdr.cutoff}{Deprecated form of fdr_cutoff}

\item{p.cutoff}{Deprecated form of p_cutoff}
}
\value{
Data frame with total significant genes for each variable at various FDR cutoffs
}
\description{
Summarise number of significant genes at various FDR cutoffs. Can split by up/down fold change as well.
}
\examples{
# Run kimma model
model_results <- kmFit(dat = example.voom,
      kin = example.kin,
      run_lme = TRUE, run_lmerel=TRUE, run_contrast=TRUE,
      subset_genes = c("ENSG00000250479","ENSG00000250510","ENSG00000255823"),
      model = "~ virus + asthma + (1|ptID)")

# Or extract limma results
# design <- model.matrix(~ virus + asthma, data = example.voom$targets)
# fit <- limma::eBayes(limma::lmFit(example.voom$E, design))
# model_results <- extract_lmFit(design = design, fit = fit)

# Summarise results
summarise_kmFit(fdr = model_results$lmerel, fdr_cutoff = c(0.01, 0.5),
                FCgroup = TRUE)
summarise_kmFit(fdr = model_results$lme.contrast, fdr_cutoff = c(0.01, 0.5),
                FCgroup = FALSE)

#No significant genes. No run
## summarise_kmFit(fdr = model_results$lmerel, fdr_cutoff = c(0.001))

}
