% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_lmFit.R
\name{summarise_lmFit}
\alias{summarise_lmFit}
\alias{summarize_lmFit}
\title{Summarise lmFit FDR results}
\usage{
summarise_lmFit(
  fdr,
  fdr_cutoff = c(0.05, 0.1, 0.2, 0.3, 0.4, 0.5),
  p_cutoff = NULL,
  FCgroup = FALSE,
  intercept = FALSE,
  fdr.cutoff = NULL,
  p.cutoff = NULL
)

summarize_lmFit(
  fdr,
  fdr_cutoff = c(0.05, 0.1, 0.2, 0.3, 0.4, 0.5),
  p_cutoff = NULL,
  FCgroup = FALSE,
  intercept = FALSE,
  fdr.cutoff = NULL,
  p.cutoff = NULL
)
}
\arguments{
\item{fdr}{data.frame output by kimma::extract_lmFit( ). Specify the model such as fdr$lm}

\item{fdr_cutoff}{numeric vector of FDR cutoffs to summarise at}

\item{p_cutoff}{numeric vector of P-value cutoffs to summarise at. No FDR summary given if p_cutoff is provided}

\item{FCgroup}{logical if should separate summary by up/down fold change groups}

\item{intercept}{logical if should include intercept variable in summary}

\item{fdr.cutoff}{Deprecated form of fdr_cutoff}

\item{p.cutoff}{Deprecated form of p_cutoff}
}
\value{
Data frame with total significant genes for each variable at various FDR cutoffs
}
\description{
Summarise number of significant genes at various FDR cutoffs. Can split by up/down fold change as well.
}
\examples{
# Run limma model
design <- model.matrix(~ virus, data = example.voom$targets)
fit <- limma::eBayes(limma::lmFit(example.voom$E, design))

## Get results
model_results <- extract_lmFit(design = design, fit = fit)

# Summarise results
summarise_lmFit(fdr = model_results$lm, fdr_cutoff = c(0.05, 0.5), FCgroup = TRUE)

# Contrasts model
design <- model.matrix(~ 0 + virus, data = example.voom$targets)
fit <- limma::lmFit(example.voom$E, design)
contrast.mat <- limma::makeContrasts(virusHRV-virusnone, levels = design)
fit <- limma::eBayes(limma::contrasts.fit(fit, contrast.mat))

## Get results
model_results <- extract_lmFit(design = design, fit = fit,
                               contrast_mat = contrast.mat)

# Summarise results
summarise_lmFit(fdr = model_results$lm.contrast,
    fdr_cutoff = c(0.05, 0.5), FCgroup = FALSE)
## No signif results (Not run error)
# summarise_lmFit(fdr = model_results$lm.contrast,
#     fdr_cutoff = 0.0001, FCgroup = FALSE)

}
