% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lmFit.R
\name{extract_lmFit}
\alias{extract_lmFit}
\title{Extract lmFit model results}
\usage{
extract_lmFit(
  design,
  fit,
  contrast.mat = NULL,
  dat.genes = NULL,
  name.genes = "geneName"
)
}
\arguments{
\item{design}{model matrix output by model.matrix( )}

\item{fit}{MArrayLM model fit output by limma::eBayes( )}

\item{contrast.mat}{contrast matrix output by limma::makeContrasts( ). NOTE: When using constrasts, the result will not exactly match extract_kmFit due to limma's naming of contrast levels as variableLEVEL}

\item{dat.genes}{data frame with additional gene annotations. Optional. If not provided, the fit object is also checked for gene annotation information.}

\item{name.genes}{character for variable name in dat.genes that matches gene names in fit}
}
\value{
List with data frames. One for model fit (sigma) and one for significance for all variable and genes. Variables names as in limma::topTable( )
}
\description{
Extract model fit and significance for all individual variables and/or contrasts in a limma model
}
\examples{
# Run limma model
design <- model.matrix(~ virus, data = example.voom$targets)
fit <- limma::eBayes(limma::lmFit(example.voom$E, design))

## Get results
fdr <- extract_lmFit(design = design, fit = fit)
## Get results and add gene annotations
fdr <- extract_lmFit(design = design, fit = fit,
                        dat.genes = example.voom$genes, name.genes = "geneName")

# Run limma contrasts model
design <- model.matrix(~ 0 + virus, data = example.voom$targets)
fit <- limma::lmFit(example.voom$E, design)
contrast.mat <- limma::makeContrasts(virusHRV-virusnone, levels = design)
fit <- limma::eBayes(limma::contrasts.fit(fit, contrast.mat))

## Get contrast results
fdr <- extract_lmFit(design = design, fit = fit, contrast.mat = contrast.mat)
}
