% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rare.R
\name{filter_rare}
\alias{filter_rare}
\title{Filter rare and low abundance genes}
\usage{
filter_rare(
  dat,
  min.CPM,
  gene.var = "geneName",
  min.sample = NULL,
  min.pct = NULL
)
}
\arguments{
\item{dat}{DGEList output by edgeR::DEGList( )}

\item{min.CPM}{numeric minimum counts per million (CPM)}

\item{gene.var}{character name for column with gene names in dat$genes that matches names in expression data dat$E. Default "geneName"}

\item{min.sample}{numeric minimum number of samples}

\item{min.pct}{numeric minimum percent of samples (0-100)}
}
\value{
DGEList
}
\description{
Filter genes at a specified minimum counts per million (CPM) in a minmum number or percent of total samples.
}
\examples{
filter_rare(dat = dat.example, min.CPM = 0.1, min.sample = 3)
filter_rare(dat = dat.example, min.CPM = 0.1, min.pct = 10)
}
