% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kimma_cleaning.R
\name{kimma_cleaning}
\alias{kimma_cleaning}
\title{kmFit data cleaning}
\usage{
kimma_cleaning(
  dat = NULL,
  kin = NULL,
  patientID = "ptID",
  libraryID = "libID",
  counts = NULL,
  meta = NULL,
  genes = NULL,
  weights = NULL,
  subset_var = NULL,
  subset_lvl = NULL,
  subset_genes = NULL,
  model_lm = NULL,
  genotype_name = NULL,
  run_lmerel = FALSE
)
}
\arguments{
\item{dat}{EList object output by voom( ). Contains counts (dat$E), meta (dat$targets), and genes (dat$genes).}

\item{kin}{Matrix with pairwise kinship values between individuals. Must be numeric with rownames.}

\item{patientID}{Character of variable name to match dat$targets to kinship row and column names.}

\item{libraryID}{Character of variable name to match dat$targets to dat$E colnames

Alternate data if not using EList object}

\item{counts}{Matrix of normalized expression. Rows are genes, columns are libraries.}

\item{meta}{Matrix or data frame of sample and individual metadata.}

\item{genes}{Matrix or data frame of gene metadata.}

\item{weights}{Matrix or data frame of gene specific weights

Subset data (optional)}

\item{subset_var}{Character list of variable name(s) to filter data by.}

\item{subset_lvl}{Character list of variable value(s) or level(s) to filter data to. Must match order of subset_var}

\item{subset_genes}{Character vector of genes to include in models.}

\item{model_lm}{Character vector of simple linear model version of model provided}

\item{genotype_name}{Character string. Used internally for kmFit_eQTL}
}
\value{
Data frame formatted for use in kmFit
}
\description{
Data
}
\keyword{internal}
