% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kimma_lm.R
\name{kimma_lm}
\alias{kimma_lm}
\title{Run kimma linear model}
\usage{
kimma_lm(model_lm, to_model_gene, gene, use_weights, metrics)
}
\arguments{
\item{model_lm}{Character model created in kmFit}

\item{to_model_gene}{Data frame formatted in kmFit, subset to gene of interest}

\item{gene}{Character of gene to model}

\item{use_weights}{Logical if gene specific weights should be used in model. Default is FALSE}

\item{metrics}{Logical if should calculate model fit metrics such as AIC, BIC, R-squared. Default is FALSE}
}
\value{
Linear model results data frame for 1 gene
}
\description{
Run kimma linear model
}
\keyword{internal}
