% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kimma_lm.R
\name{kimma_lm}
\alias{kimma_lm}
\title{Run kimma linear model}
\usage{
kimma_lm(model.lm, to.model.gene, gene)
}
\arguments{
\item{model.lm}{Character model created in kmFit}

\item{to.model.gene}{Data frame formatted in kmFit, subset to gene of interest}

\item{gene}{Character of gene to model}
}
\value{
Linear model results data frame for 1 gene
}
\description{
Run kimma linear model
}
\keyword{internal}
