% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kimma_lmerel.R
\name{kimma_lmerel}
\alias{kimma_lmerel}
\title{Run kimma linear mixed effects model with kinship}
\usage{
kimma_lmerel(
  model_lme,
  to_model_gene,
  gene,
  kin_subset,
  use_weights,
  patientID,
  metrics
)
}
\arguments{
\item{model_lme}{Character model created in kmFit}

\item{to_model_gene}{Data frame formatted in kmFit, subset to gene of interest}

\item{gene}{Character of gene to model}

\item{kin_subset}{Pairwise kinship matrix}

\item{use_weights}{Logical if gene specific weights should be used in model. Default is FALSE}

\item{patientID}{Character of variable name to match dat$targets to kinship row and column names.}

\item{metrics}{Logical if should calculate model fit metrics such as AIC, BIC, R-squared. Default is FALSE}
}
\value{
Linear mixed effect with kinship results data frame for 1 gene
}
\description{
Run kimma linear mixed effects model with kinship
}
\keyword{internal}
