% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmFit.R
\name{kmFit}
\alias{kmFit}
\title{Linear mixed effects models with kinship for RNA-seq}
\usage{
kmFit(
  dat = NULL,
  kin = NULL,
  patientID = "ptID",
  libraryID = "libID",
  counts = NULL,
  meta = NULL,
  genes = NULL,
  subset.var = NULL,
  subset.lvl = NULL,
  subset.genes = NULL,
  model,
  run.lm = FALSE,
  run.lme = FALSE,
  run.lmekin = FALSE,
  run.contrast = FALSE,
  contrast.var = NULL,
  processors = NULL,
  p.method = "BH"
)
}
\arguments{
\item{dat}{EList object output by voom( ). Contains counts (dat$E), meta (dat$targets), and genes (dat$genes).}

\item{kin}{Matrix with pairwise kinship values between individuals. Must be numeric with rownames.}

\item{patientID}{Character of variable name to match dat$targets to kinship row and column names.}

\item{libraryID}{Character of variable name to match dat$targets to dat$E colnames}

\item{counts}{Matrix of normalized expression. Rows are genes, columns are libraries.}

\item{meta}{Matrix or data frame of sample and individual metadata.}

\item{genes}{Matrix or data frame of gene metadata.}

\item{subset.var}{Character list of variable name(s) to filter data by.}

\item{subset.lvl}{Character list of variable value(s) or level(s) to filter data to. Must match order of subset.var}

\item{subset.genes}{Character vector of genes to include in models.}

\item{model}{Character vector of model starting with ~ Should include (1|patientID) if mixed effects will be run}

\item{run.lm}{Logical if should run lm model without kinship}

\item{run.lme}{Logical if should run lme model without kinship}

\item{run.lmekin}{Logical if should run lmekin model with kinship}

\item{run.contrast}{Logical if should run pairwise contrasts. If no matrix provided, all possible pairwise comparisons are completed.}

\item{contrast.var}{Character vector of variable in model to run contrasts of. Interaction terms must be specified as "var1:var2". If NULL (default), all contrasts for all variables in the model are run}

\item{processors}{Numeric processors to run in parallel. Default is 2 less than the total available}

\item{p.method}{Character of FDR adjustment method. Values as in p.adjust( )}
}
\value{
Dataframe with model fit and significance for each gene
}
\description{
Run lmekin and corresponding lm or lme without kinship of gene expression in RNA-seq data
}
\examples{
# All samples and all genes
kmFit(dat = example.voom,
      patientID = "donorID", libraryID = "libID",
      kin = example.kin, run.lmekin = TRUE,
      model = "~ virus + (1|donorID)")

# Subset samples and genes
kmFit(dat = example.voom,
      patientID = "donorID", libraryID = "libID",
      run.lme = TRUE,
      subset.var = list("asthma"), subset.lvl = list(c("asthma")),
      subset.genes = c("ENSG00000250479","ENSG00000250510","ENSG00000255823"),
      model = "~ virus + (1|donorID)")
# Pairwise contrasts
kmFit(dat = example.voom,
      patientID = "donorID", libraryID = "libID",
      run.lme = TRUE, run.contrast = TRUE,
      subset.genes = c("ENSG00000250479","ENSG00000250510","ENSG00000255823"),
      model = "~ virus+asthma * median_cv_coverage + (1|donorID)",
      contrast.var=c("virus","asthma:median_cv_coverage"))

kmFit(dat = example.voom, kin = example.kin,
      patientID = "donorID", libraryID = "libID",
      run.lmekin = TRUE, run.contrast = TRUE,
      subset.genes = c("ENSG00000250479","ENSG00000250510","ENSG00000255823"),
      model = "~ virus*asthma + (1|donorID)",
      contrast.var=c("virus","virus:asthma"))

# Model with failed genes
kmFit(dat = example.voom,
      patientID = "donorID", libraryID = "libID",
      kin = example.kin, run.lmekin = TRUE, run.lm = TRUE,
      subset.genes = c("ENSG00000250479","ENSG00000250510","ENSG00000255823"),
      model = "~ virus*asthma + lib.size + norm.factors + median_cv_coverage + donorID+(1|donorID)")
}
