% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmFit_contrast.R
\name{kmFit_contrast}
\alias{kmFit_contrast}
\title{Run pairwise model comparisons with emmeans}
\usage{
kmFit_contrast(fit, contrast.vars, contrast.list)
}
\arguments{
\item{fit}{model fit from lm( ) or lmer( )}

\item{contrast.vars}{character vector of variables to test}

\item{contrast.list}{contrast matrix formatted as list}
}
\value{
data frame with contrast model results
}
\description{
Run pairwise model comparisons with emmeans
}
\keyword{internal}
