% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmFit_contrast.R
\name{kmFit_contrast}
\alias{kmFit_contrast}
\title{Run pairwise model comparisons with emmeans}
\usage{
kmFit_contrast(fit, contrast.var, to.model.gene, metrics)
}
\arguments{
\item{fit}{model fit from lm( ) or lmer( )}

\item{contrast.var}{Character vector of variable in model to run contrasts of}

\item{to.model.gene}{Formatted data from kimma_cleaning( ), subset to gene of interest}

\item{metrics}{Logical if should calculate model fit metrics such as AIC, BIC, R-squared. Default is FALSE}
}
\value{
data frame with contrast model results
}
\description{
Run pairwise model comparisons with emmeans
}
\keyword{internal}
