% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmFit_eQTL.R
\name{kmFit_eQTL}
\alias{kmFit_eQTL}
\title{eQTL linear mixed effects models with kinship}
\usage{
kmFit_eQTL(
  dat.snp = NULL,
  dat.map = NULL,
  geneID = "gene",
  genotypeID = "snpID",
  dat = NULL,
  kin = NULL,
  patientID = "ptID",
  libraryID = "libID",
  counts = NULL,
  meta = NULL,
  genes = NULL,
  weights = NULL,
  subset.var = NULL,
  subset.lvl = NULL,
  subset.genes = NULL,
  model,
  use.weights = FALSE,
  run.lm = FALSE,
  run.lme = FALSE,
  run.lmerel = FALSE,
  metrics = FALSE,
  run.lmekin = NULL,
  run.contrast = FALSE,
  contrast.var = NULL,
  processors = NULL,
  p.method = "BH"
)
}
\arguments{
\item{dat.snp}{Data frame containing 0,1,2 genotypes. Can be numeric or character/factor depending on desired analysis. If running contrasdts, must be character/factor for pairwise comparisons to run. Rows are donors and columns are SNP.}

\item{dat.map}{Data frame mapping geneID to genotypeID. If provided, genotypes will be modeled against only the genes they are paired with in the map. If not provided, genotypes will be modeled against all genes in dat$E / counts}

\item{geneID}{Character variable name to match dat.snp dat with gene identifiers. Values should match rownames in express data (dat$E or counts) Default it "gene"}

\item{genotypeID}{Character variable name to match dat.snp dat with SNP identifiers. Default it "snpID"}

\item{dat}{EList object output by voom( ). Contains counts (dat$E), meta (dat$targets), and genes (dat$genes).}

\item{kin}{Matrix with pairwise kinship values between individuals. Must be numeric with rownames.}

\item{patientID}{Character of variable name to match dat$targets to kinship row and column names.}

\item{libraryID}{Character of variable name to match dat$targets to dat$E colnames}

\item{counts}{Matrix of normalized expression. Rows are genes, columns are libraries.}

\item{meta}{Matrix or data frame of sample and individual metadata.}

\item{genes}{Matrix or data frame of gene metadata.}

\item{weights}{Matrix of data frame of gene specific weights. Usually calculated with limma::voomWithQualityWeights().}

\item{subset.var}{Character list of variable name(s) to filter data by.}

\item{subset.lvl}{Character list of variable value(s) or level(s) to filter data to. Must match order of subset.var}

\item{subset.genes}{Character vector of genes to include in models.}

\item{model}{Character vector of model starting with ~ Should include (1|patientID) if mixed effects will be run. NOTE THAT FOR EQTL, the model MUST include genotype as a place holder for the genotype data. For example "~ genotype + (1|ptID)"

Same as kmFit( )}

\item{use.weights}{Logical if gene specific weights should be used in model. Default is FALSE}

\item{run.lm}{Logical if should run lm model without kinship}

\item{run.lme}{Logical if should run lme model without kinship}

\item{run.lmerel}{Logical if should run lmerel model with kinship}

\item{metrics}{Logical if should calculate model fit metrics such as AIC, BIC, R-squared. Default is FALSE}

\item{run.lmekin}{Depreciated. Please use run.lmerel}

\item{run.contrast}{Logical if should run pairwise contrasts. If no matrix provided, all possible pairwise comparisons are completed.}

\item{contrast.var}{Character vector of variable in model to run contrasts of. Interaction terms must be specified as "var1:var2". If NULL (default), all contrasts for all variables in the model are run}

\item{processors}{Numeric processors to run in parallel. Default is 2 less than the total available}

\item{p.method}{Character of FDR adjustment method. Values as in p.adjust( )}
}
\value{
List of data frames including
   - lm/lme/lmerel: model estimates and significance
   - *.contrast: model estimates and significance for pairwise contrasts with variables in the original model
   - *.fit: model fit metrics such as sigma, AIC, BIC, R-squared (optional with metrics paramater)
   - *.error: error messages for genes that failed model fitting
}
\description{
Run lmerel and corresponding lm or lme without kinship of gene expression in RNA-seq data against associated genotypes
}
\examples{
#Determine data size
nSNP <- 4
nSamp <- length(unique(example.voom$targets$ptID))
#Create random genotype data
snp <- data.frame(matrix(sample(c(0,1,2), nSNP*nSamp, replace=TRUE),
                         nrow=nSamp)) \%>\%
  #Add SNP names
  dplyr::mutate(ptID = unique(example.voom$targets$ptID)) \%>\%
  dplyr::select(ptID, dplyr::everything())
colnames(snp) <- c("ptID", paste0("rs",1:nSNP))

#Create gene to SNP map
 map <- data.frame(snpID = paste0("rs",1:nSNP),
                 gene = rep(example.voom$genes$geneName[1:2], nSNP/2))

#Run SNP interaction model
eQTL_result <- kmFit_eQTL(dat.snp = snp, dat.map = map,
                          dat = example.voom, kin = example.kin,
                          run.lmerel = TRUE, metrics=TRUE,
                          run.contrast = TRUE, contrast.var="virus:genotype",
                          model = "~ virus*genotype + (1|ptID)")

#Run SNP as factors with contrasts
snp_fct <- snp \%>\%
 dplyr::mutate_if(is.numeric, ~as.factor(.))

eQTL_result2 <- kmFit_eQTL(dat.snp = snp_fct, dat.map = map,
                           dat = example.voom, kin = example.kin,
                           run.lmerel = TRUE, run.contrast = TRUE, metrics=TRUE,
                           model = "~ genotype + (1|ptID)")
}
